moved some of the misc_utils into command_utils
