+++ utils/misc_utils.py
@@ -2,30 +2,7 @@ import ipaddress
 import subprocess
 from shutil import which
 
-__all__ = ["human_time", "human_size", "sanitize_hostname", "sudo_open", "parse_version",
-           "noop_command", "failed_to_load_command"]
-
-from types import SimpleNamespace
-
-from dt_shell import DTCommandAbs, dtslogger
-
-
-class NoOpCommand(DTCommandAbs):
-
-    @staticmethod
-    def command(shell, args):
-        pass
-
-
-class FailedToLoadCommand(NoOpCommand):
-
-    @staticmethod
-    def command(shell, args):
-        dtslogger.warning("This command was not loaded")
-
-
-noop_command = SimpleNamespace(DTCommand=NoOpCommand)
-failed_to_load_command = SimpleNamespace(DTCommand=FailedToLoadCommand)
+__all__ = ["human_time", "human_size", "sanitize_hostname", "sudo_open", "parse_version"]
 
 
 def human_time(time_secs, compact=False):
